+++ eslint-plugin/tests/lib/rules/no-import-all-from-library.test.js
@@ -17,12 +17,14 @@
  */
 
 "use strict";
+
 //------------------------------------------------------------------------------
 // Requirements
 //------------------------------------------------------------------------------
 
-const rule = require("../../../lib/rules/limit-db-query-results"),
-  RuleTester = require("eslint").RuleTester;
+const rule = require("../../../lib/rules/no-import-all-from-library");
+const { RuleTester } = require("eslint");
+const { describe, it } = require("node:test");
 
 //------------------------------------------------------------------------------
 // Tests
@@ -34,46 +36,54 @@ const ruleTester = new RuleTester({
     sourceType: "module",
   },
 });
-
 const expectedError = {
-  messageId: "LimitTheNumberOfReturns",
-  type: "Literal",
+  messageId: "ShouldNotImportAllFromLibrary",
 };
 
-ruleTester.run("limit-db-query-results", rule, {
+const tests = {
   valid: [
     `
-      sqlClient.query("SELECT id, name, email FROM customers LIMIT 10;");
+    import isEmpty from 'lodash/isEmpty';
     `,
     `
-      sqlClient.query("SELECT TOP 5 * FROM products;");
+    import orderBy from 'lodash/orderBy';
     `,
     `
-      sqlClient.query("SELECT id, name, email FROM customers WHERE id = 1");
+    import { orderBy } from 'lodash-es';
     `,
     `
-      sqlClient.query("SELECT * FROM orders FETCH FIRST 20 ROWS ONLY");
-    `,
-    `
-      sqlClient.query("WITH numbered_customers AS (SELECT *, ROW_NUMBER() OVER (ORDER BY customer_id) AS row_num FROM customers) SELECT * FROM numbered_customers WHERE row_num <= 50");
-    `,
-    `
-      console.log("SELECT id, name, email FROM customers WHERE id = 1");
+    import map from 'underscore/modules/map.js';
     `,
   ],
 
   invalid: [
     {
-      code: `sqlClient.query("SELECT * FROM bikes");`,
+      code: "import lodash from 'lodash';",
       errors: [expectedError],
     },
     {
-      code: `sqlClient.run("SELECT id, departure, arrival FROM flights");`,
+      code: "import * as lodash from 'lodash';",
       errors: [expectedError],
     },
     {
-      code: `sqlClient.execute("SELECT * FROM cars");`,
+      code: "import * as lodash from 'lodash-es';",
+      errors: [expectedError],
+    },
+    {
+      code: "import someLib from 'some-lib';",
+      options: [{ notAllowedLibraries: ["some-lib"] }],
+      errors: [expectedError],
+    },
+    {
+      code: "import * as someLib from 'some-lib';",
+      options: [{ importByNamespaceNotAllowedLibraries: ["some-lib"] }],
       errors: [expectedError],
     },
   ],
+};
+
+describe("no-import-all-from-library", () => {
+  it("no-import-all-from-library", () => {
+    ruleTester.run("no-import-all-from-library", rule, tests);
+  });
 });