+++ eslint-plugin/tests/lib/rules/no-multiple-style-changes.test.js
@@ -23,7 +23,8 @@
 //------------------------------------------------------------------------------
 
 const rule = require("../../../lib/rules/no-multiple-style-changes");
-const RuleTester = require("eslint").RuleTester;
+const { RuleTester } = require("eslint");
+const { describe, it } = require("node:test");
 
 //------------------------------------------------------------------------------
 // Tests
@@ -37,10 +38,9 @@ const ruleTester = new RuleTester({
 });
 const expectedError = {
   messageId: "UseClassInstead",
-  type: "AssignmentExpression",
 };
 
-ruleTester.run("no-multiple-style-changes", rule, {
+const tests = {
   valid: [
     {
       code: 'element.style.height = "800px";',
@@ -115,4 +115,10 @@ ruleTester.run("no-multiple-style-changes", rule, {
       errors: [expectedError],
     },
   ],
+};
+
+describe("no-multiple-style-changes", () => {
+  it("no-multiple-style-changes", () => {
+    ruleTester.run("no-multiple-style-changes", rule, tests);
+  });
 });