+++ eslint-plugin/tests/lib/rules/prefer-collections-with-pagination.test.js
@@ -23,7 +23,8 @@
 //------------------------------------------------------------------------------
 
 const rule = require("../../../lib/rules/prefer-collections-with-pagination");
-const RuleTester = require("eslint").RuleTester;
+const { RuleTester } = require("eslint");
+const { describe, it } = require("node:test");
 
 //------------------------------------------------------------------------------
 // Tests
@@ -37,14 +38,12 @@ const ruleTester = new RuleTester({
 
 const expectedArrayError = {
   messageId: "PreferReturnCollectionsWithPagination",
-  type: "TSArrayType",
 };
 const expectedReferenceError = {
   messageId: "PreferReturnCollectionsWithPagination",
-  type: "TSTypeReference",
 };
 
-ruleTester.run("prefer-collections-with-pagination", rule, {
+const tests = {
   valid: [
     `
     @Controller()
@@ -124,4 +123,10 @@ ruleTester.run("prefer-collections-with-pagination", rule, {
       errors: [expectedReferenceError],
     },
   ],
+};
+
+describe("prefer-collections-with-pagination", () => {
+  it("prefer-collections-with-pagination", () => {
+    ruleTester.run("prefer-collections-with-pagination", rule, tests);
+  });
 });